add user api, upload image

+++ echo-api/src/posts/posts.service.ts
@@ -1,9 +1,10 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { UsersService } from '../users/users.service';
-import { AUTH_MESSAGES, POST_MESSAGES } from '../messages';
+import { COMMON_MESSAGES, POST_MESSAGES } from '../common/messages';
+import { AuthRequest } from '../auth/interfaces/auth-request.interface';
 
 @Injectable()
 export class PostsService {
@@ -11,24 +12,27 @@ export class PostsService {
     private readonly userService: UsersService
   ) {}
 
-  async create(createPostDto: CreatePostDto) {
-    const user = await this.userService.findByUsername(createPostDto.authorId);
-    if (!user) throw new NotFoundException(AUTH_MESSAGES.ERROR.NOT_FOUND_USERNAME);
+  async create(createPostDto: CreatePostDto, auth: AuthRequest) {
+    await this.userService.findOne(auth.id);
 
-    const post = await this.prisma.post.create({
+    if (createPostDto.rootPostId) await this.findOne(createPostDto.rootPostId);
+    if (createPostDto.parentPostId) await this.findOne(createPostDto.parentPostId);
+
+    const createdPost = await this.prisma.post.create({
       data: {
         ...createPostDto,
+        authorId: auth.id,
       }
     });
 
     if (!createPostDto.rootPostId) {
       return await this.prisma.post.update({
-        where: { id: post.id },
-        data: { rootPostId: post.rootPostId },
+        where: { id: createdPost.id },
+        data: { rootPostId: createdPost.id },
       });
     }
 
-    return post;
+    return { messages: POST_MESSAGES.SUCCESS.CREATE_POST, post: createdPost };
   }
 
   async findAll() {
@@ -41,20 +45,27 @@ export class PostsService {
     return post;
   }
 
-  async update(id: number, updatePostDto: UpdatePostDto) {
-    await this.findOne(id);
+  async update(id: number, updatePostDto: UpdatePostDto, auth: AuthRequest) {
+    const user = await this.userService.findOne(auth.id);
+    const post = await this.findOne(id);
+    if (user.id !== post.authorId) throw new UnauthorizedException(COMMON_MESSAGES.ERROR.UNAUTHORIZED);
 
-    return await this.prisma.post.update({
+    const updatedPost = await this.prisma.post.update({
       where: { id },
       data: {
         ...updatePostDto,
       }
     });
+
+    return { messages: POST_MESSAGES.SUCCESS.UPDATE_POST, post: updatedPost };
   }
 
-  async remove(id: number) {
-    const post = await this.findOne(id);
+  async remove(id: number, auth: AuthRequest) {
+    const user = await this.userService.findOne(auth.id);
+    const removedPost = await this.findOne(id);
+    if (user.id !== removedPost.authorId) throw new UnauthorizedException(COMMON_MESSAGES.ERROR.UNAUTHORIZED);
+
     await this.prisma.post.delete({ where: { id }});
-    return post;
+    return { messages: POST_MESSAGES.SUCCESS.DELETE_POST, post: removedPost };
   }
 }
